created router files and moved playground and review routes
