+++ views/playgrounds/show.ejs
@@ -40,7 +40,7 @@
             Rating: <%= review.rating %>
           </h5>
           <p class="card-text">Review: <%= review.body %></p>
-          <a href="/playgrounds/<%= playground._id %>/review/<%= review._id %>/edit" class="card-link btn btn-sm btn-info">Edit</a>
+          <!-- <a href="/playgrounds/<%= playground._id %>/review/<%= review._id %>/edit" class="card-link btn btn-sm btn-info">Edit</a> -->
           <form class="d-inline" action="/playgrounds/<%= playground._id %>/review/<%= review._id %>?_method=DELETE" method="POST">
             <button class="card-link btn btn-sm btn-danger">Delete</button>
           </form>
